Widen app_timeout from uint16 to uint32 to support timeouts >65535s
Values like inactivity-timeout 86400 (24h) exceeded the uint16 max
(65535), silently dropping to 0. Widened the entire chain:
- BPF: app_value.timeout, pkt_meta.app_timeout, session_value.app_timeout
- Go: AppValue.Timeout, SetApplication() signature, compiler cap removed
- DPDK: shared_mem.h structs (session, app_value, pkt_meta)
- Session struct: app_timeout (u32) placed in existing padding gap
  between is_reverse and created — zero struct growth
- Added "timeout" as synonym for "inactivity-timeout" in app parsing

+++ pkg/dataplane/dataplane.go
@@ -83,7 +83,7 @@ type DataPlane interface {
 	ClearAddressMembership() error
 
 	// Application
-	SetApplication(protocol uint8, dstPort uint16, appID uint32, timeout uint16, algType uint8, srcPortLow, srcPortHigh uint16) error
+	SetApplication(protocol uint8, dstPort uint16, appID uint32, timeout uint32, algType uint8, srcPortLow, srcPortHigh uint16) error
 	ClearApplications() error
 
 	// Sessions
